pyproject.toml
[tool.poetry]
name = "umbrella"
version = "0.1.0"
description = "a discord bot with minecraft server & aws integration"
authors = ["hyunwoo312 <hyunwoojames@gmail.com>"]
license = "MIT"
readme = "README.md"


[tool.poetry.dependencies]
python = "3.11.*"
discord-py = "2.3.2"
coloredlogs = "15.0.1"
pydantic = "2.3.0"
pydantic-settings = "2.0.3"


[tool.poetry.group.dev.dependencies]
awscli = "1.29.*"
boto3 = "1.28.*"

black = "23.7.0"
flake8 = "6.1.0"
pytest = "7.4.0"
pytest-cov = "4.1.0"

pre-commit = "3.0.2"
mypy = "1.5.1"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.taskipy.tasks]
start = "python3 -m umbrella"
